Fixes after code review comments
