+++ src/main/java/org/concordion/internal/command/strategies/Default.java
@@ -1,15 +1,11 @@
 package org.concordion.internal.command.strategies;
 
-import org.concordion.api.CommandCall;
-import org.concordion.api.Evaluator;
-import org.concordion.api.Result;
-import org.concordion.api.ResultRecorder;
+import org.concordion.api.*;
 import org.concordion.api.listener.VerifyRowsListener;
 import org.concordion.internal.Row;
-import org.concordion.internal.command.VerifyRowsCommand;
 import org.concordion.internal.util.Announcer;
 
-public class Default extends VerifyRowsCommand.VerificationStrategy {
+public class Default extends VerifyRowsStrategy {
     public Default(CommandCall commandCall, Evaluator evaluator, ResultRecorder resultRecorder,
                    Announcer<VerifyRowsListener> listeners, String loopVariableName, Iterable<Object> actualRows) {
         super(commandCall, evaluator, resultRecorder, listeners, loopVariableName, actualRows);
