+++ src/test/resources/spec/concordion/command/verifyRows/strategies/Default.html
@@ -174,7 +174,7 @@
         </tr>
 
         <tr>
-            <td>One surpulus row</td>
+            <td>One surplus row</td>
             <td>
                 <pre class="html">
 &lt;table concordion:verifyRows="#user : users" concordion:verificationStrategy="Default"&gt;
